Struct google_containeranalysis1_beta1::Occurrence

pub struct Occurrence {
    pub kind: Option<String>,
    pub installation: Option<GrafeasV1beta1PackageDetails>,
    pub name: Option<String>,
    pub vulnerability: Option<GrafeasV1beta1VulnerabilityDetails>,
    pub update_time: Option<String>,
    pub discovered: Option<GrafeasV1beta1DiscoveryDetails>,
    pub attestation: Option<Details>,
    pub build: Option<GrafeasV1beta1BuildDetails>,
    pub deployment: Option<GrafeasV1beta1DeploymentDetails>,
    pub remediation: Option<String>,
    pub resource: Option<ResourceType>,
    pub create_time: Option<String>,
    pub derived_image: Option<GrafeasV1beta1ImageDetails>,
    pub note_name: Option<String>,
}

An instance of an analysis type that has been found on a resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

occurrences create projects (request|response)

occurrences get projects (response)

occurrences patch projects (request|response)

Fields
kind: Option<String>

Output only. This explicitly denotes which of the occurrence details are specified. This field can be used as a filter in list requests.

installation: Option<GrafeasV1beta1PackageDetails>

Describes the installation of a package on the linked resource.

name: Option<String>

Output only. The name of the occurrence in the form of projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID].

vulnerability: Option<GrafeasV1beta1VulnerabilityDetails>

Describes a security vulnerability.

update_time: Option<String>

Output only. The time this occurrence was last updated.

discovered: Option<GrafeasV1beta1DiscoveryDetails>

Describes when a resource was discovered.

attestation: Option<Details>

Describes an attestation of an artifact.

build: Option<GrafeasV1beta1BuildDetails>

Describes a verifiable build.

deployment: Option<GrafeasV1beta1DeploymentDetails>

Describes the deployment of an artifact on a runtime.

remediation: Option<String>

A description of actions that can be taken to remedy the note.

resource: Option<ResourceType>

Required. Immutable. The resource for which the occurrence applies.

create_time: Option<String>

Output only. The time this occurrence was created.

derived_image: Option<GrafeasV1beta1ImageDetails>

Describes how this resource derives from the basis in the associated note.

note_name: Option<String>

Required. Immutable. The analysis note associated with this occurrence, in the form of projects/[PROVIDER_ID]/notes/[NOTE_ID]. This field can be used as a filter in list requests.