Struct google_containeranalysis1_beta1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the ContainerAnalysis hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_containeranalysis1_beta1 as containeranalysis1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use containeranalysis1_beta1::ContainerAnalysis;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ContainerAnalysis::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `notes_batch_create(...)`, `notes_create(...)`, `notes_delete(...)`, `notes_get(...)`, `notes_get_iam_policy(...)`, `notes_list(...)`, `notes_occurrences_list(...)`, `notes_patch(...)`, `notes_set_iam_policy(...)`, `notes_test_iam_permissions(...)`, `occurrences_batch_create(...)`, `occurrences_create(...)`, `occurrences_delete(...)`, `occurrences_get(...)`, `occurrences_get_iam_policy(...)`, `occurrences_get_notes(...)`, `occurrences_get_vulnerability_summary(...)`, `occurrences_list(...)`, `occurrences_patch(...)`, `occurrences_set_iam_policy(...)`, `occurrences_test_iam_permissions(...)`, `scan_configs_get(...)`, `scan_configs_list(...)` and `scan_configs_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn scan_configs_list(
&self,
parent: &str
) -> ProjectScanConfigListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists scan configurations for the specified project.

Arguments

parent - The name of the project to list scan configurations for in the form of projects/[PROJECT_ID].

pub fn notes_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectNoteTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the permissions that a caller has on the specified note or occurrence. Requires list permission on the project (for example, containeranalysis.notes.list).

The resource takes the format projects/[PROJECT_ID]/notes/[NOTE_ID] for notes and projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID] for occurrences.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn occurrences_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectOccurrenceSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified note or occurrence. Requires containeranalysis.notes.setIamPolicy or containeranalysis.occurrences.setIamPolicy permission if the resource is a note or an occurrence, respectively.

The resource takes the format projects/[PROJECT_ID]/notes/[NOTE_ID] for notes and projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID] for occurrences.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn occurrences_batch_create(
&self,
request: BatchCreateOccurrencesRequest,
parent: &str
) -> ProjectOccurrenceBatchCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates new occurrences in batch.

Arguments

request - No description provided.

parent - The name of the project in the form of projects/[PROJECT_ID], under which the occurrences are to be created.

pub fn notes_get(&self, name: &str) -> ProjectNoteGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified note.

Arguments

name - The name of the note in the form of projects/[PROVIDER_ID]/notes/[NOTE_ID].

pub fn occurrences_get(&self, name: &str) -> ProjectOccurrenceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified occurrence.

Arguments

name - The name of the occurrence in the form of projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID].

pub fn notes_patch(
&self,
request: Note,
name: &str
) -> ProjectNotePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified note.

Arguments

request - No description provided.

name - The name of the note in the form of projects/[PROVIDER_ID]/notes/[NOTE_ID].

pub fn notes_occurrences_list(
&self,
name: &str
) -> ProjectNoteOccurrenceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists occurrences referencing the specified note. Provider projects can use this method to get all occurrences across consumer projects referencing the specified note.

Arguments

name - The name of the note to list occurrences for in the form of projects/[PROVIDER_ID]/notes/[NOTE_ID].

pub fn notes_list(&self, parent: &str) -> ProjectNoteListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists notes for the specified project.

Arguments

parent - The name of the project to list notes for in the form of projects/[PROJECT_ID].

pub fn occurrences_patch(
&self,
request: Occurrence,
name: &str
) -> ProjectOccurrencePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified occurrence.

Arguments

request - No description provided.

name - The name of the occurrence in the form of projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID].

pub fn occurrences_get_vulnerability_summary(
&self,
parent: &str
) -> ProjectOccurrenceGetVulnerabilitySummaryCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a summary of the number and severity of occurrences.

Arguments

parent - The name of the project to get a vulnerability summary for in the form of projects/[PROJECT_ID].

pub fn occurrences_create(
&self,
request: Occurrence,
parent: &str
) -> ProjectOccurrenceCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new occurrence.

Arguments

request - No description provided.

parent - The name of the project in the form of projects/[PROJECT_ID], under which the occurrence is to be created.

pub fn notes_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectNoteGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a note or an occurrence resource. Requires containeranalysis.notes.setIamPolicy or containeranalysis.occurrences.setIamPolicy permission if the resource is a note or occurrence, respectively.

The resource takes the format projects/[PROJECT_ID]/notes/[NOTE_ID] for notes and projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID] for occurrences.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn notes_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectNoteSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified note or occurrence. Requires containeranalysis.notes.setIamPolicy or containeranalysis.occurrences.setIamPolicy permission if the resource is a note or an occurrence, respectively.

The resource takes the format projects/[PROJECT_ID]/notes/[NOTE_ID] for notes and projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID] for occurrences.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn occurrences_get_notes(
&self,
name: &str
) -> ProjectOccurrenceGetNoteCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the note attached to the specified occurrence. Consumer projects can use this method to get a note that belongs to a provider project.

Arguments

name - The name of the occurrence in the form of projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID].

pub fn scan_configs_update(
&self,
request: ScanConfig,
name: &str
) -> ProjectScanConfigUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified scan configuration.

Arguments

request - No description provided.

name - The name of the scan configuration in the form of projects/[PROJECT_ID]/scanConfigs/[SCAN_CONFIG_ID].

pub fn occurrences_list(
&self,
parent: &str
) -> ProjectOccurrenceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists occurrences for the specified project.

Arguments

parent - The name of the project to list occurrences for in the form of projects/[PROJECT_ID].

pub fn occurrences_delete(
&self,
name: &str
) -> ProjectOccurrenceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified occurrence. For example, use this method to delete an occurrence when the occurrence is no longer applicable for the given resource.

Arguments

name - The name of the occurrence in the form of projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID].

pub fn notes_batch_create(
&self,
request: BatchCreateNotesRequest,
parent: &str
) -> ProjectNoteBatchCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates new notes in batch.

Arguments

request - No description provided.

parent - The name of the project in the form of projects/[PROJECT_ID], under which the notes are to be created.

pub fn notes_delete(&self, name: &str) -> ProjectNoteDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified note.

Arguments

name - The name of the note in the form of projects/[PROVIDER_ID]/notes/[NOTE_ID].

pub fn scan_configs_get(&self, name: &str) -> ProjectScanConfigGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified scan configuration.

Arguments

name - The name of the scan configuration in the form of projects/[PROJECT_ID]/scanConfigs/[SCAN_CONFIG_ID].

pub fn occurrences_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectOccurrenceGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a note or an occurrence resource. Requires containeranalysis.notes.setIamPolicy or containeranalysis.occurrences.setIamPolicy permission if the resource is a note or occurrence, respectively.

The resource takes the format projects/[PROJECT_ID]/notes/[NOTE_ID] for notes and projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID] for occurrences.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn notes_create(
&self,
request: Note,
parent: &str
) -> ProjectNoteCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new note.

Arguments

request - No description provided.

parent - The name of the project in the form of projects/[PROJECT_ID], under which the note is to be created.

pub fn occurrences_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectOccurrenceTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the permissions that a caller has on the specified note or occurrence. Requires list permission on the project (for example, containeranalysis.notes.list).

The resource takes the format projects/[PROJECT_ID]/notes/[NOTE_ID] for notes and projects/[PROJECT_ID]/occurrences/[OCCURRENCE_ID] for occurrences.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.